+++ tools/tpm2_sign.c
@@ -28,20 +28,18 @@ struct tpm_sign_ctx {
     char *outFilePath;
     BYTE *msg;
     UINT16 length;
-    char *inMsgFileName;
+    char *input_file;
     tpm2_convert_sig_fmt sig_format;
 
     struct {
-        UINT8 m : 1;
+        UINT8 d : 1;
         UINT8 t : 1;
         UINT8 o : 1;
-        UINT8 D : 1;
     } flags;
 };
 
 static tpm_sign_ctx ctx = {
         .halg = TPM2_ALG_SHA1,
-        .digest = NULL,
         .sig_scheme = TPM2_ALG_NULL
 };
 
@@ -50,18 +48,7 @@ static tool_rc sign_and_save(ESYS_CONTEXT *ectx) {
     TPMT_SIGNATURE *signature;
     bool result;
 
-    tool_rc rc = tool_rc_general_error;
-
-    if (!ctx.flags.D) {
-      tool_rc tmp_rc = tpm2_hash_compute_data(ectx, ctx.halg, TPM2_RH_NULL,
-              ctx.msg, ctx.length, &ctx.digest, NULL);
-      if (tmp_rc != tool_rc_success) {
-          LOG_ERR("Compute message hash failed!");
-          return tmp_rc;
-      }
-    }
-
-    rc = tpm2_sign(ectx, &ctx.signing_key.object, ctx.digest, &ctx.in_scheme,
+    tool_rc rc = tpm2_sign(ectx, &ctx.signing_key.object, ctx.digest, &ctx.in_scheme,
       &ctx.validation, &signature);
     if (rc != tool_rc_success) {
       goto out;
@@ -70,6 +57,7 @@ static tool_rc sign_and_save(ESYS_CONTEXT *ectx) {
     result = tpm2_convert_sig_save(signature, ctx.sig_format,
                 ctx.outFilePath);
     if (!result) {
+        rc = tool_rc_general_error;
         goto out;
     }
 
@@ -89,11 +77,6 @@ static tool_rc init(ESYS_CONTEXT *ectx) {
         option_fail = true;
     }
 
-    if (!ctx.flags.m && !ctx.flags.D) {
-        LOG_ERR("Expected options m or D");
-        option_fail = true;
-    }
-
     if (!ctx.flags.o) {
         LOG_ERR("Expected option o");
         option_fail = true;
@@ -103,11 +86,12 @@ static tool_rc init(ESYS_CONTEXT *ectx) {
         return tool_rc_option_error;
     }
 
-    if (ctx.flags.D && (ctx.flags.t || ctx.flags.m)) {
-        LOG_WARN("Option D provided, options m and t are ignored.");
+    if (ctx.flags.d && ctx.flags.t) {
+        LOG_WARN("When using a pre-computed digest the validation ticket"
+                " is ignored.");
     }
 
-    if (ctx.flags.D || !ctx.flags.t) {
+    if (ctx.flags.d || !ctx.flags.t) {
         ctx.validation.tag = TPM2_ST_HASHCHECK;
         ctx.validation.hierarchy = TPM2_RH_NULL;
         memset(&ctx.validation.digest, 0, sizeof(ctx.validation.digest));
@@ -123,39 +107,38 @@ static tool_rc init(ESYS_CONTEXT *ectx) {
         return rc;
     }
 
-    /*
-     * Process the msg file if needed
-     */
-    if (ctx.flags.m && !ctx.flags.D) {
-      unsigned long file_size;
-      bool result = files_get_file_size_path(ctx.inMsgFileName, &file_size);
-      if (!result) {
-          return tool_rc_general_error;
-      }
-      if (file_size == 0) {
-          LOG_ERR("The message file \"%s\" is empty!", ctx.inMsgFileName);
+    /* Process the msg file if needed */
+    if (!ctx.flags.d) {
+      FILE *input = ctx.input_file ? fopen(ctx.input_file, "rb") : stdin;
+      if (!input) {
+          LOG_ERR("Could not open file \"%s\"", ctx.input_file);
           return tool_rc_general_error;
       }
 
-      if (file_size > UINT16_MAX) {
-          LOG_ERR(
-                  "The message file \"%s\" is too large, got: %lu bytes, expected less than: %u bytes!",
-                  ctx.inMsgFileName, file_size, UINT16_MAX + 1);
-          return tool_rc_general_error;
+      rc = tpm2_hash_file(ectx, ctx.halg, TPM2_RH_NULL, input,
+              &ctx.digest, NULL);
+      if (input != stdin) {
+          fclose(input);
       }
+      if (rc != tool_rc_success) {
+          LOG_ERR("Could not hash input");
+      }
+      return rc;
+      /* we don't need to perform the digest, just read it */
+    }
 
-      ctx.msg = (BYTE*) calloc(required_argument, file_size);
-      if (!ctx.msg) {
-          LOG_ERR("oom");
-          return tool_rc_general_error;
-      }
+    /* else process it as a pre-computed digest */
+    ctx.digest = malloc(sizeof(TPM2B_DIGEST));
+    if (!ctx.digest) {
+        LOG_ERR("oom");
+        return tool_rc_general_error;
+    }
 
-      ctx.length = file_size;
-      result = files_load_bytes_from_path(ctx.inMsgFileName, ctx.msg, &ctx.length);
-      if (!result) {
-          free(ctx.msg);
-          return tool_rc_general_error;
-      }
+    ctx.digest->size = sizeof(ctx.digest->buffer);
+    bool result = files_load_bytes_from_buffer_or_file_or_stdin(NULL,
+            ctx.input_file, &ctx.digest->size, ctx.digest->buffer);
+    if (!result) {
+        return tool_rc_general_error;
     }
 
     return tool_rc_success;
@@ -170,14 +153,13 @@ static bool on_option(char key, char *value) {
     case 'p':
         ctx.signing_key.auth_str = value;
         break;
-    case 'g': {
+    case 'g':
         ctx.halg = tpm2_alg_util_from_optarg(value, tpm2_alg_util_flags_hash);
         if (ctx.halg == TPM2_ALG_ERROR) {
             LOG_ERR("Could not convert to number or lookup algorithm, got: \"%s\"",
                     value);
             return false;
         }
-    }
         break;
     case 's': {
         ctx.sig_scheme = tpm2_alg_util_from_optarg(value, tpm2_alg_util_flags_sig);
@@ -187,19 +169,8 @@ static bool on_option(char key, char *value) {
         }
     }
         break;
-    case 'd': {
-        ctx.digest = malloc(sizeof(TPM2B_DIGEST));
-        ctx.digest->size = sizeof(TPM2B_DIGEST);
-        if (!files_load_bytes_from_path(value, ctx.digest->buffer, &ctx.digest->size)) {
-            LOG_ERR("Could not load digest from file \"%s\"!", value);
-            return false;
-        }
-        ctx.flags.D = 1;
-    }
-        break;
-    case 'm':
-        ctx.inMsgFileName = value;
-        ctx.flags.m = 1;
+    case 'd':
+        ctx.flags.d = 1;
         break;
     case 't': {
         bool result = files_load_validation(value, &ctx.validation);
@@ -209,10 +180,9 @@ static bool on_option(char key, char *value) {
         ctx.flags.t = 1;
     }
         break;
-    case 'o': {
+    case 'o':
         ctx.outFilePath = value;
         ctx.flags.o = 1;
-    }
         break;
     case 'f':
         ctx.sig_format = tpm2_convert_sig_fmt_from_optarg(value);
@@ -226,22 +196,33 @@ static bool on_option(char key, char *value) {
     return true;
 }
 
+static bool on_args(int argc, char *argv[]) {
+
+    if (argc != 1) {
+        LOG_ERR("Expected one input file, got: %d", argc);
+        return false;
+    }
+
+    ctx.input_file = argv[0];
+
+    return true;
+}
+
 bool tpm2_tool_onstart(tpm2_options **opts) {
 
     static const struct option topts[] = {
       { "auth",                 required_argument, NULL, 'p' },
       { "hash-algorithm",       required_argument, NULL, 'g' },
       { "scheme",               required_argument, NULL, 's' },
-      { "message",              required_argument, NULL, 'm' },
-      { "digest",               required_argument, NULL, 'd' },
+      { "digest",               no_argument,       NULL, 'd' },
       { "signature",            required_argument, NULL, 'o' },
       { "ticket",               required_argument, NULL, 't' },
       { "key-context",          required_argument, NULL, 'c' },
       { "format",               required_argument, NULL, 'f' }
     };
 
-    *opts = tpm2_options_new("p:g:m:d:t:o:c:f:s:", ARRAY_LEN(topts), topts,
-                             on_option, NULL, 0);
+    *opts = tpm2_options_new("p:g:dt:o:c:f:s:", ARRAY_LEN(topts), topts,
+                             on_option, on_args, 0);
 
     return *opts != NULL;
 }